+++ src/userdoc/select2/574_noform.php
@@ -1,6 +1,6 @@
 <?php // Issue #574: multiselect without form should send all selected elements
 $this->widget(
-    'bootstrap.widgets.TbSelect2',
+    'booster.widgets.TbSelect2',
     array(
         'name' => 'group_id_list',
         'data' => array('RU' => 'Russian Federation', 'CA' => 'Canada', 'US' => 'United States of America', 'GB' => 'Great Britain'),
@@ -12,7 +12,7 @@ $this->widget(
 );
 echo CHtml::endForm();
 $this->widget(
-    'bootstrap.widgets.TbButton',
+    'booster.widgets.TbButton',
     array(
         'label' => 'Click on me with Developer Tools opened!',
         'htmlOptions' => array(