+++ src/userdoc/select2/empty_data.php
@@ -1,6 +1,6 @@
 <?php // Initially empty TbSelect2 (data is null)
 $this->widget(
-    'bootstrap.widgets.TbSelect2',
+    'booster.widgets.TbSelect2',
     array(
         'name' => 'emptydata',
         'data' => null,